Cleanup the old code
